Deprecate old week page, initializing new week page to work with JournalLayout
