+++ src/app/week/deprecated/week.scss
@@ -1,0 +1,15 @@
+.week-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  main {
+    display: flex;
+    flex-direction: row;
+    margin-bottom: 50px;
+    @media only screen and (max-width: 600px) {
+      flex-direction: column;
+      margin-bottom: 0px;
+    }
+  }
+}
