+++ src/app/week/week.css
@@ -1,17 +1,4 @@
+@import url("https://fonts.googleapis.com/css2?family=Assistant:wght@200..800&family=Tenor+Sans&display=swap");
 .week-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-.week-container main {
-  display: flex;
-  flex-direction: row;
-  margin-bottom: 50px;
-}
-@media only screen and (max-width: 600px) {
-  .week-container main {
-    flex-direction: column;
-    margin-bottom: 0px;
-  }
+  background-color: #f1daad;
 }/*# sourceMappingURL=week.css.map */
